refactor(copilot_task_submit): split tests_orchestration.rs (#1266)
tests_orchestration.rs 528 -> 267 LOC; new tests_orchestration_extra.rs.
